+++ CHANGELOG.md
@@ -7,6 +7,8 @@ project adheres to [Semantic Versioning][SemVer].
 ### Added
 - Add experimental support for modules and include a `mod_example` with the
   source code.  The APIs aren't documented yet and may change in the future.
+- Include `mod_stats_influx`, a module for logging STUN/TURN events/statistics
+  to InfluxDB, contributed by Marc Schink.
 
 ## [1.1.0] - 2020-07-22
 ### Added